+++ contracts/interfaces/treasury/ITreasuryExecution.sol
@@ -2,18 +2,16 @@
 pragma solidity ^0.8.9;
 
 interface ITreasuryExecution {
-
   event ProposalExecuted(
-    uint indexed proposalId,
+    uint256 indexed proposalId,
     address indexed destination,
-    uint indexed value,
+    uint256 indexed value,
     bytes callData
   );
 
-  function executeProposal(uint proposalId) external returns(bool result);
+  function executeProposal(uint256 proposalId) external returns (bool result);
 
   //function createSubTreasuryType0() external;
 
   //function createSubTreasuryType1() external;
-
 }
