+++ contracts/interfaces/treasury/ITreasuryVoting.sol
@@ -2,56 +2,42 @@
 pragma solidity ^0.8.9;
 
 interface ITreasuryVoting {
-
-  event TreasuryProposalCreated(
-    uint indexed proposalId,
-    uint indexed deadlineTimestamp
-  );
+  event TreasuryProposalCreated(uint256 indexed proposalId, uint256 indexed deadlineTimestamp);
 
   event TreasuryProposalAccepted(
-    uint indexed proposalId,
+    uint256 indexed proposalId,
     address indexed destination,
-    uint indexed value,
+    uint256 indexed value,
     bytes callData
   );
 
   event TreasuryProposalRejected(
-    uint indexed proposalId,
+    uint256 indexed proposalId,
     address indexed destination,
-    uint indexed value,
+    uint256 indexed value,
     bytes callData
   );
 
   function createTreasuryProposal(
     address destination,
-    uint value,
+    uint256 value,
     bytes calldata callData,
     uint128 deadlineTimestamp
-  ) external returns(uint);
+  ) external returns (uint256);
 
   function createSeveralTreasuryProposals(
     address[] calldata destinations,
-    uint[] calldata values,
+    uint256[] calldata values,
     bytes[] calldata callDatas,
     uint128[] calldata deadlineTimestamps
-  ) external returns(uint[] memory);
+  ) external returns (uint256[] memory);
 
-  function voteForOneTreasuryProposal(
-    uint proposalId,
-    bool vote
-  ) external;
+  function voteForOneTreasuryProposal(uint256 proposalId, bool vote) external;
 
-  function voteForSeveralTreasuryProposals(
-    uint[] calldata proposalsIds,
-    bool[] calldata votes
-  ) external;
+  function voteForSeveralTreasuryProposals(uint256[] calldata proposalsIds, bool[] calldata votes)
+    external;
 
-  function acceptOrRejectProposal(
-    uint proposalId
-  ) external;
+  function acceptOrRejectProposal(uint256 proposalId) external;
 
-  function acceptOrRejectSeveralProposals(
-    uint[] calldata proposalIds
-  ) external;
-
+  function acceptOrRejectSeveralProposals(uint256[] calldata proposalIds) external;
 }
